# Conflicts:
#	HoloLens Spring 2019 LTS Support/Assets/Scenes/PAXScene.unity
#	HoloLens Spring 2019 LTS Support/Assets/Scripts/BaseEnemy.cs
#	HoloLens Spring 2019 LTS Support/Assets/Scripts/UpgradeManager.cs
#	HoloLens Spring 2019 LTS Support/Assets/Scripts/WaveSpawner.cs
#	HoloLens Spring 2019 LTS Support/ProjectSettings/TagManager.asset

+++ HoloLens Spring 2019 LTS Support/Assets/Scripts/GameManager.cs
@@ -28,7 +28,7 @@ public class GameManager : MonoBehaviour
     public bool Spawn;
 
     [Header("Coins")]
-    public int coins;
+    public float coins;
 
     [Header("Spawner")]
     public GameObject[] enemyPrefab;
